Address change in inherited member type (ROOT5 vs ROOT6)
The type of `TListIter::fCursor` and `TListIter::fCurCursor` (inherited by
`StFileIter`) changes from `TObjLink*` to `std::shared_ptr<TObjLink>` in ROOT5
and ROOT6 respectively.

In order to solve the problem with assignment operator we introduce a couple of
overloads one of which is selected based on the actual type of the respective
data members in `TListIter`.